Created a new microservice structure
