+++ .gitignore
@@ -1,2 +1,2 @@
 .env
-_pycache_/*
+__pycache__/*
