+++ app/core/constants/appConstants.py
